requirements.txt
streamlit>=1.30.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.4.0
duckduckgo-search>=4.1.1
newspaper3k>=0.2.8
google-search-results>=2.4.2
lxml[html_clean]>=4.9.3 